Inline `FlightPhase` class
